Fix: Correctly handle multiple socket connections
The previous implementation in scripts/socket.js used a single variable
to store the last connected socket, which would not work correctly for
multiple concurrent users.

This change replaces the single variable with a Map (`connectedSockets`)
to store all connected sockets, keyed by their `socket.id`. Sockets are
added on connection and removed on disconnection. The exported accessor
function has been renamed from `get_socket` to `get_sockets` for
clarity and now returns the entire Map.

Refactor: Move API routes to api.routes.js

For better organization and to address a TODO comment, several API-like
routes have been moved from routes/routes.js to a new file,
routes/api.routes.js.

The moved routes include:
- GET /weather/:latlon
- POST /alert
- GET /deviceLatest/:esp
- POST /saveProfile
- GET /data/:options

The main server file, sbqc_serv.js, has been updated to mount these
routes under the /api prefix. Necessary dependencies and environment
variable access have been included in the new api.routes.js file.

Fix: Improve error handling and UX on registration form

Previously, errors during user registration (validation failures or
email already exists) in routes/login.routes.js would return a raw
HTTP 400 response.

This change modifies the behavior to re-render the
views/partials/login/register.ejs template with an error message and
repopulate your input (name and email). The register.ejs
template has been updated to display these error messages and
pre-fill form fields. This resolves a TODO comment.

Refactor: Use constant for login redirect path

In routes/login.routes.js, the redirect path '../settings' after a
successful login was hardcoded. This has been changed to use a
LOGIN_SUCCESS_REDIRECT_PATH constant defined at the top of the file
for better maintainability. This addresses a TODO comment.

Docs: Clarify auto-creation of MongoDB 'boot' collection

A TODO comment in sbqc_serv.js questioned the creation of the 'boot'
collection. This change replaces the TODO with a comment explaining
that MongoDB's insertOne operation automatically creates the collection
if it doesn't exist, unless specific creation options are needed.

Chore: Remove outdated TODO for root visit logging

A TODO comment in routes/routes.js suggested logging root path visits
to the database. This functionality was already implemented (logging
to the 'server' collection). The outdated TODO comment has been removed.
