+++ routes/routes.js
@@ -8,7 +8,7 @@ const getMqtt = require('../scripts/mqttServer').getClient
 const Tools = require('nodetools')
 
 
-const mailman = require('../public/js/mailman')
+// const mailman = require('../public/js/mailman'); // This line is removed as mailman is used in api.routes.js
 let counter = require('../scripts/visitorCount')
 const sysmon = new (require('../scripts/systemMonitor'))()
 console.log("SysInfo: ", sysmon.getinfo().data)
@@ -16,7 +16,7 @@ console.log("CPU: ", sysmon.getinfo().cpus.length)
 
 const apiUrl =process.env.DATA_API_URL + (process.env.DATA_API_PORT ? ":" + process.env.DATA_API_PORT : "") //let dAPIUrl = "https://data.specialblend.ca"
 //const apiUrl = "http://" + process.env.DATA_API_IP + ":" + process.env.DATA_API_PORT
-//const mqttUrl = "ws://" + process.env.DATA_API_URL + ":" + process.env.MQTT_PORT 
+//const mqttUrl = "ws://" + process.env.DATA_API_URL + ":" + process.env.MQTT_PORT
 const mqttWSUrl = process.env.MQTT_SERVER_WS
 const mqttinfo = JSON.stringify({url: mqttWSUrl, user: process.env.USER, pass: process.env.PASS })
 
@@ -31,10 +31,10 @@ const getUserLogs = async (req, res, next) => {
     // let limit = Number(req.query.limit) || 10
     let { skip = 0,  sort = 'desc', source = 'userLogs' } = req.query
     skip = parseInt(skip) || 0
-    
+
 
     skip = skip < 0 ? 0 : skip;
-   
+
     const logsdb =  req.app.locals.collections[source]
     console.log('Getting logs from DB namespace', logsdb.namespace)
 
@@ -65,7 +65,7 @@ const createUserLog = async (req, res, next) => {
         res.json(createdLog)
     }
     catch(err) {console.log(err); next() }
-        
+
     } else {
         res.status(422)
         res.json({
@@ -77,12 +77,11 @@ const createUserLog = async (req, res, next) => {
 
 function requestLog(req) {
 
-    //  TODO: send dans BD ces infos pour un checkin log de qui vient sur root /
-    let client = req.headers['user-agent'];        
-    let content = req.headers['Content-Type'];     
-    let authorize = req.headers['Authorization'];   
-    let origin = req.headers['host'];              
-    let ip = req.socket.remoteAddress;     
+    let client = req.headers['user-agent'];
+    let content = req.headers['Content-Type'];
+    let authorize = req.headers['Authorization'];
+    let origin = req.headers['host'];
+    let ip = req.socket.remoteAddress;
 
     let queryParams = req.query; // Query parameters
     let path = req.path; // Path of the request URL
@@ -92,7 +91,7 @@ function requestLog(req) {
     let originalUrl = req.originalUrl; // Original URL of the request
     let cookies = req.cookies; // Cookies sent by the client (if any)   const cookieParser = require('cookie-parser');      app.use(cookieParser());
 
-    
+
     const log = {
         logType: 'checkin',
         client: client ? client.toString().trim() : 'none',
@@ -124,14 +123,14 @@ function requestLog(req) {
 ////   free routes
 
 router.get("/", async (req, res) => {
-   
+
 
     let count = await counter.increaseCount()
 
     const log = requestLog(req)
-      
+
     const logsdb =  req.app.locals.collections.server;
-        
+
     try{
         const createdLog = await logsdb.insertOne(log)
         console.log(`Log document was inserted with the _id: ${createdLog.insertedId}`)
@@ -159,33 +158,33 @@ router.get('/cams',  (req, res) => {  res.render('cams')  })
 router.get('/dashboard', async (req, res) => {
     console.log('Getting registered Esp32')
     const registered = await esp32.getRegistered()
-    if(registered == null) { 
-        console.log('Could not fetch devices list. Is DataAPI online?') 
-        res.render('index',    { name: req.session.email }) 
+    if(registered == null) {
+        console.log('Could not fetch devices list. Is DataAPI online?')
+        res.render('index',    { name: req.session.email })
     } else {
         console.log(registered.map((dev) => id = dev.id ))
         res.render('dashboard', { title: "Dashboard", menuId: 'home', hitCount: await counter.getCount(), collectionInfo: req.app.locals.collectionInfo, regDevices: registered })
-    } 
-    
+    }
+
 })
 
 
-router.get('/iot',  async (req, res) => 
-{ 
+router.get('/iot',  async (req, res) =>
+{
     console.log('Getting registered Esp32')
     const registered = await esp32.getRegistered()
-    if(registered == null) { 
-        console.log('Could not fetch devices list. Is DataAPI online?') 
-        res.render('index',    { name: req.session.email }) 
+    if(registered == null) {
+        console.log('Could not fetch devices list. Is DataAPI online?')
+        res.render('index',    { name: req.session.email })
     } else {
         console.log(registered.map((dev) => id = dev.id ))
         res.render('iot', { mqttinfo: mqttinfo, regDevices: registered, dataApiStatus: esp32.dataApiStatus  })
-    } 
+    }
 })
 
-    
-router.get('/earth', async (req, res) => {   
-    res.render('earth')  
+
+router.get('/earth', async (req, res) => {
+    res.render('earth')
 })
 
 router.get('/natureCode', (req, res) => {
@@ -227,10 +226,10 @@ router.get('/serverspec', (req,res) => {
 
 
 
-router.get('/device',  async (req, res) => 
+router.get('/device',  async (req, res) =>
 {
     try{
-        const registered = await esp32.getRegistered()  
+        const registered = await esp32.getRegistered()
         if(!registered.length) {
             console.log('No devices registered yet!!!!! Cannot display device page, redirecting....')
             res.redirect('/iot')
@@ -238,16 +237,16 @@ router.get('/device',  async (req, res) =>
             let selectedDevice = req.session.selectedDevice ? req.session.selectedDevice : registered[0].id  //  default on 1st device if none is saved in session
             selectedDevice = req.query.deviceID ? req.query.deviceID : selectedDevice // selection from query superceed saved session
             console.log('Fetching Alarms for: ' + selectedDevice)
-        
+
             const response2 = await fetch(apiUrl + "/alarms")
             const alarmList = await response2.json()
-    
+
             let selDevice
             registered.forEach(device =>{ if(device.id == selectedDevice) {  selDevice = device }  })
             console.log('Selected Device:', selDevice.id, selDevice.config[0])
-    
+
             res.render('device', { mqttinfo: mqttinfo, devices: registered, device: selDevice, alarmList: alarmList, apiUrl: apiUrl, iGrowUrl: req.protocol + '://' + req.get('host')  })
-        }  
+        }
     } catch(err) {
             res.render('error', { mqttinfo: mqttinfo, devices: devices, selected: selectedDevice, device: selDevice, alarmList: alarmList, apiUrl: apiUrl, iGrowUrl: req.protocol + '://' + req.get('host')  })
     }
@@ -255,14 +254,14 @@ router.get('/device',  async (req, res) =>
 
 
 
-router.get('/graphs',  async (req, res) => 
-{ 
+router.get('/graphs',  async (req, res) =>
+{
     const response = await fetch(apiUrl + "/devices")
     const result = await response.json()
     const list = result.data
-    let selectedDevice = req.session.selectedDevice ? req.session.selectedDevice : list[0].id  // req.query.deviceID ? req.query.deviceID : list[0]  
+    let selectedDevice = req.session.selectedDevice ? req.session.selectedDevice : list[0].id  // req.query.deviceID ? req.query.deviceID : list[0]
     console.log('loading graf: ', selectedDevice )
-    
+
 
     const registered = await esp32.getRegistered()
     const devices = { list, registered }
@@ -272,11 +271,11 @@ router.get('/graphs',  async (req, res) =>
 
 
 
-router.post('/selectDevice', async (req, res) => 
+router.post('/selectDevice', async (req, res) =>
 {
     req.session.selectedDevice = req.body.selected
     console.log('Receiving selection: ' , req.body.selected)
-    req.session.save(async (err) => { 
+    req.session.save(async (err) => {
         if(err) console.log('Session error: ', err)
        // console.log(req.session)
         res.redirect('/graphs')
@@ -298,7 +297,7 @@ router.get('/database',  async (req, res) =>
 
 //  Session validation & logged in routes    -  User will have a saved Session if logged in
 
-const hasSessionID = (req, res, next) => 
+const hasSessionID = (req, res, next) =>
 {
     console.log('Session: ', req.session)
     if (!req.session.userToken) {
@@ -310,7 +309,7 @@ const hasSessionID = (req, res, next) =>
 
 
 
-router.get('/settings',  hasSessionID,  async (req, res) => 
+router.get('/settings',  hasSessionID,  async (req, res) =>
 {
     const response = await fetch(apiUrl + "/users")
     const result = await response.json()
@@ -324,211 +323,13 @@ router.get('/settings',  hasSessionID,  async (req, res) =>
     const result3 = await response3.json()
     const alarms = result3.data
 
-    res.render('settings', {users: users, devices: devices, alarms: alarms})  
+    res.render('settings', {users: users, devices: devices, alarms: alarms})
 })
 
+// API routes are now in routes/api.routes.js (This comment replaces the old TODO and the routes themselves)
 
 
-
-
-
-
-
-
-
-
-//  API extra   --   TODO:  doit surement etre dans des routes séparées
-
-router.get('/weather/:latlon', async (req, res) => {
-
-    const [lat, lon] = req.params.latlon.split(',');
-   const weatherURL = `https://api.openweathermap.org/data/2.5/weather?lat=${lat}&lon=${lon}&units=metric&APPID=${process.env.WEATHER_API_KEY}`;
-  
-    const aq_url = `https://api.openaq.org/v3/locations?coordinates=${lat},${lon}&radius=5000`;
-
-    console.log(lat, lon);
-
-    try {
-        const weather_response = await fetch(weatherURL);
-       // console.log(weather_response)
-        if (!weather_response.ok) {
-            throw new Error(`Failed to fetch weather data: ${weather_response.statusText}`);
-        }
-        const weather = await weather_response.json();
-        console.log('Weather data:', weather);
-
-        const aq_response = await fetch(aq_url, {
-            headers: {
-                'X-API-Key': process.env.OPENAQ_API_KEY
-            }
-        });
-        if (!aq_response.ok) {
-            const errorText = await aq_response.text();
-            throw new Error(`Failed to fetch air quality data: ${aq_response.statusText} - ${errorText}`);
-        }
-        const aq_data = await aq_response.json();
-        //console.log(aq_data);
-
-        const sensor =  aq_data.results
-        if (Array.isArray(sensor)) {
-            sensor.sort((a, b) => new Date(b.datetimeLast) - new Date(a.datetimeLast));
-        }
-
-        const sensorAQ = sensor[0]
-        console.log('Sensor:', sensorAQ.id)
-
-        const sensor_url = `https://api.openaq.org/v3/sensors/${sensorAQ.id}`;
-        const sensor_response = await fetch(sensor_url, {
-            headers: {
-                'X-API-Key': process.env.OPENAQ_API_KEY
-            }
-        });
-        const sensor_data = await sensor_response.json();
-        console.log('Sensor data:', sensor_data)
-
-        const data = {
-            weather: weather,
-            air_quality: sensor_data
-        };
-        res.json(data);
-    } catch (error) {
-        console.error(error);
-        res.status(500).json({ error: 'Failed to fetch weather or air quality data' });
-    }
-});
-
-
-router.post('/alert', async (req, res) => {
-
-    console.log('post to Alert:')
-    //console.log(req.body)
-   // let dat = JSON.stringify(req.body)
-   // console.log(dat)
-
-    const alert = req.body
-    const dest = alert.dest
-    const msg = alert.msg
-    const image64 = alert.image64
-    console.log(dest, msg);
-    
-    mailman.sendEmail(dest, msg, image64)
-   
-})
-
-
-
-
-router.get('/deviceLatest/:esp',  async (req, res) => 
-{
-    let option = {
-        method: 'GET',
-        headers: {
-            'auth-token': req.session.userToken
-        }
-    }
-
-    try {
-        const response = await fetch(apiUrl + "/heartbeats/senderLatest/" + req.params.esp, option)
-        const respData = await response.json()
-        const data = respData.data[0]
-        //console.log(data)
-
-        if (!data) {
-            res.json({ status: "error", message: 'Could not get data, no latest post', data: null  })
-            //return res.status(400).send(message);
-        }
-        else {
-
-           /* let now =  new moment()
-            let stamp =  new moment(data.time).format('YYYY-MM-DD HH:mm:ss') 
-            let duration = new moment.duration(now.diff(stamp)).asHours();
-     
-            data.lastConnect = duration
-
-            if(data.wifi != -100) {
-                 if(duration > 0.05)  
-                 {
-                    console.log('Device disconnected !!!!!!!!!!!!!!!!!!!!!!!!!!!!!!')
-                    data.wifi   = -100
-                    delete data._id
-                    let dat = JSON.stringify(data)
-                    console.log(dat)
-                    let mq = getMqtt()
-                    mq.publish('esp32/alive/'+req.params.esp, dat)  //  server sends a mqtt post on behalf of esp to log a last wifi -100 signal in db.
-                 }
-             }*/
-
-            res.json({ status: "success", message: "Latest post retreived", data: data  })
-        }
-    }
-    catch (err) {
-        console.error(err)
-    }
-})
-
-router.post('/saveProfile', async (req, res) => {
-    const { profileName, config } = req.body;
-
-    const profileData = {
-        profileName,
-        config
-    };
-
-    const option = {
-        method: 'PATCH',
-        headers: {
-            'Content-Type': 'application/json',
-            'auth-token': req.session.userToken // Assuming you have a session-based auth token
-        },
-        body: JSON.stringify(profileData)
-    };
-
-    try {
-        const response = await fetch(apiUrl + "/profile/" + profileData.profileName, option);
-        const result = await response.json();
-
-        if (response.ok) {
-            res.send('Profile saved successfully!');
-        } else {
-            console.error('Error saving profile:', result);
-            res.status(response.status).send('Error saving profile: ' + result.message);
-        }
-    } catch (err) {
-        console.error('Error connecting to Data API:', err);
-        res.status(500).send('Error connecting to Data API: ' + err.message);
-    }
-});
-
-
-router.get('/data/:options',  async (req, res) => 
-{
-    const options = req.params.options.split(',')
-    const samplingRatio = options[0]
-    const espID = options[1]
-    const dateFrom = options[2]
-    const ratio = Number(samplingRatio)
-    console.log({ ratio, espID, dateFrom })
-
-
-    req.session.selectedDevice = espID
-
-    let option = { method: 'GET', headers: { 'auth-token': req.session.userToken  }    }
-
-    try {
-        const response = await fetch(apiUrl + "/heartbeats/data/" + samplingRatio + "," + espID + ',' + dateFrom, option)
-        const respData = await response.json()
-        const data = respData.data
-        res.json(data)
-    }
-    catch (err) {
-        console.error(err)
-        return res.status(400).send("Could not get data");
-    }
-})
-
-
-
-router.post('/set_io', (req, res) => 
+router.post('/set_io', (req, res) =>
 {
     let msg = 'esp32/' + req.body.sender + '/io/' + req.body.io_id + '/' + (req.body.io_state === 'ON' ? 'on' : 'off')
     console.log('Setting IO: ' + msg)
@@ -541,8 +342,8 @@ router.post('/set_io', (req, res) =>
 
 
 
-router.route('/alarms/setAlarm').post(async (req, res) => 
-{ 
+router.route('/alarms/setAlarm').post(async (req, res) =>
+{
     console.log('post received: Set_alarm')
     //console.log(JSON.stringify(req.body))
 
@@ -550,20 +351,20 @@ router.route('/alarms/setAlarm').post(async (req, res) =>
     als.espID =   req.body.device_id //'ESP_35030'  //  'ESP_15605'    ESP_35030
     als.io = req.body.io_id
     als.tStart = req.body.tStart //moment(req.body.tStart).format('YYYY-MM-DD HH:MM:SS')
-    als.tStop = req.body.tStop 
+    als.tStop = req.body.tStop
 
     let option = {
         method: 'POST',
         headers: {
             'auth-token': req.session.userToken ,
-            'Content-type': 'application/json'   
+            'Content-type': 'application/json'
         },
         body: JSON.stringify(als)
     }
     try {
         const response = await fetch( apiUrl + "/alarms", option)
         const data = await response.json()
-     
+
         if (Tools.isObjEmpty(data)) {
             const message = "Error saving alarm";
             console.log(message)
@@ -571,7 +372,7 @@ router.route('/alarms/setAlarm').post(async (req, res) =>
         }
         else {  //  send new alarm to already connected ESP.  Non connected ESP will receive the alarm at next boot.
             let mq = getMqtt()
-            let topic = 'esp32/' + als.espID + '/io' 
+            let topic = 'esp32/' + als.espID + '/io'
             let startTime = moment(als.tStart).local().format('HH:mm:ss')
             let stopTime = moment(als.tStop).local().format('HH:mm:ss')
             mq.publish('esp32/' + als.espID + '/io/sunrise', als.io + ":" + startTime)
